pyproject.toml
[project]
name = "3d-camera"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.4",
    "pygame-ce>=2.5.7",
    "pyside6>=6.11.0",
]

[dependency-groups]
dev = []
